test(charges): consolidate custom currency coverage
