+++ openmeter/billing/charges/service/usagebased_costbasis_test.go
@@ -38,7 +38,7 @@ func (s *UsageBasedCostBasisCreateSuite) SetupSuite() {
 }
 
 func (s *UsageBasedCostBasisCreateSuite) TearDownTest() {
-	s.setCustomCurrencyEnabled(false)
+	s.setUsageBasedCustomCurrencyEnabled(false)
 	s.BaseSuite.TearDownTest()
 }
 
@@ -106,7 +106,7 @@ func (s *UsageBasedCostBasisCreateSuite) TestCreatePersistsManualPinnedAndDynami
 		FiatCurrency: fiatCurrency,
 	})
 
-	s.setCustomCurrencyEnabled(true)
+	s.setUsageBasedCustomCurrencyEnabled(true)
 	created, err := s.Charges.usageBasedService.Create(ctx, usagebased.CreateInput{
 		Namespace: namespace,
 		Intents: []usagebased.Intent{
@@ -191,7 +191,7 @@ func (s *UsageBasedCostBasisCreateSuite) TestSetResolvedDynamicCostBasisIsRetryS
 	dynamicIntent := costbasis.NewIntent(costbasis.DynamicIntent{
 		FiatCurrency: s.newFiatCurrency("USD"),
 	})
-	s.setCustomCurrencyEnabled(true)
+	s.setUsageBasedCustomCurrencyEnabled(true)
 	created, err := s.Charges.usageBasedService.Create(ctx, usagebased.CreateInput{
 		Namespace: namespace,
 		Intents: []usagebased.Intent{
@@ -277,7 +277,7 @@ func (s *UsageBasedCostBasisCreateSuite) TestDynamicCostBasisResolvesWhenChargeB
 	dynamicIntent := costbasis.NewIntent(costbasis.DynamicIntent{
 		FiatCurrency: s.newFiatCurrency("USD"),
 	})
-	s.setCustomCurrencyEnabled(true)
+	s.setUsageBasedCustomCurrencyEnabled(true)
 	clock.FreezeTime(activationAt)
 	defer clock.UnFreeze()
 	created, err := s.Charges.usageBasedService.Create(ctx, usagebased.CreateInput{
@@ -353,7 +353,7 @@ func (s *UsageBasedCostBasisCreateSuite) TestPinnedCostBasisMustMatchCurrencyAnd
 		{name: "fiat currency", costBasisID: eurCostBasis.ID, errorText: "currency cost basis fiat currency mismatch"},
 	}
 
-	s.setCustomCurrencyEnabled(true)
+	s.setUsageBasedCustomCurrencyEnabled(true)
 	for _, test := range tests {
 		s.Run(test.name, func() {
 			// given:
@@ -398,7 +398,7 @@ func (s *UsageBasedCostBasisCreateSuite) TestCreateRollsBackCostBasesWhenChargeC
 		Rate:         alpacadecimal.NewFromInt(2),
 	})
 
-	s.setCustomCurrencyEnabled(true)
+	s.setUsageBasedCustomCurrencyEnabled(true)
 	_, err := s.Charges.usageBasedService.Create(ctx, usagebased.CreateInput{
 		Namespace: namespace,
 		Intents: []usagebased.Intent{
@@ -418,7 +418,7 @@ func (s *UsageBasedCostBasisCreateSuite) TestCreateWithoutCostBasisLeavesChargeR
 	customer := s.CreateTestCustomer(namespace, "usage-based-without-cost-basis")
 	currency := s.createTestCustomCurrency(ctx, namespace)
 	featureMeters := s.createFeatureMeters(ctx, namespace, "credit-only-feature")
-	s.setCustomCurrencyEnabled(true)
+	s.setUsageBasedCustomCurrencyEnabled(true)
 
 	created, err := s.Charges.usageBasedService.Create(ctx, usagebased.CreateInput{
 		Namespace: namespace,
@@ -437,13 +437,6 @@ func (s *UsageBasedCostBasisCreateSuite) TestCreateWithoutCostBasisLeavesChargeR
 	s.Require().NoError(err)
 	s.Require().Nil(chargeEntity.CostBasisID)
 	s.Require().Equal(0, s.countCostBases(namespace))
-}
-
-func (s *UsageBasedCostBasisCreateSuite) setCustomCurrencyEnabled(enabled bool) {
-	s.T().Helper()
-	enabler, ok := s.Charges.usageBasedService.(customCurrencyEnabler)
-	s.Require().True(ok)
-	s.Require().NoError(enabler.SetEnableCustomCurrency(s.T(), enabled))
 }
 
 func (s *UsageBasedCostBasisCreateSuite) countCostBases(namespace string) int {
